refactor: replace hand-rolled ULID ID types with ulid_id! macro (INC-13)
Add a declarative ulid_id! macro to modo core that generates
ULID-backed newtype IDs with standard traits (Display, FromStr,
Serialize, Deserialize, From, AsRef, etc.). Replace the hand-rolled
SessionId and JobId implementations with single macro invocations.
